--- a/pqarrow/parquet.go
+++ b/pqarrow/parquet.go
@@ -1,0 +1,27 @@
+package pqarrow
+
+import (
+	"fmt"
+
+	"github.com/apache/arrow/go/v8/arrow"
+	"github.com/apache/arrow/go/v8/arrow/scalar"
+	"github.com/segmentio/parquet-go"
+)
+
+func ArrowScalarToParquetValue(sc scalar.Scalar) (parquet.Value, error) {
+	switch s := sc.(type) {
+	case *scalar.String:
+		return parquet.ValueOf(string(s.Data())), nil
+	case *scalar.Int64:
+		return parquet.ValueOf(s.Value), nil
+	case *scalar.FixedSizeBinary:
+		width := s.Type.(*arrow.FixedSizeBinaryType).ByteWidth
+		v := [16]byte{}
+		copy(v[:width], s.Data())
+		return parquet.ValueOf(v), nil
+	case nil:
+		return parquet.Value{}, nil
+	default:
+		return parquet.Value{}, fmt.Errorf("unsupported scalar type %T", s)
+	}
+}