… db_test.go
… distinct_test.go
… pqarrow/arrow_test.go
… query/logicalplan/builder_test.go
… query/logicalplan/logicalplan_test.go
… query/logicalplan/optimize.go
… query/logicalplan/optimize_test.go
… query/physicalplan/distinct.go
… query/physicalplan/physicalplan.go
… query/physicalplan/physicalplan_test.go
… query/physicalplan/project.go
… table.go
… table_test.go