1) [Optimize distinct queries with single value per rowgroup](https://github.com/polarsignals/frostdb/commit/ce1959ed7e64d06d240400dd256315b49a546090):
Gathering distinct values of (non optional) dictionary encoded columns
is a constant operation per rowgroup when the dictionary only has a
single value. Additionally distinct projections using binary expressions
such as `timestamp > 0` can be determined without loading the row if the
column index has sufficient data. For example when the index notes that
all timestamps are between 5 and 8, then all values for `timestamp > 0`
are true, and therefore true for any distinct values in the rowgroup.

2) [pqarrow: Optimize distinct where all but one column are single value](https://github.com/polarsignals/frostdb/commit/872f3fcffed60a17989f579203c7a617e576e168):
When all but one column have only a single value, then the whole
rowgroup does not have to be scanned as the possible combination of
distinct values are the non-single column values combined with all
single value columns.