@brancz 2022-07-24
With these two optimizations profile type queries in [Parca](https://parca.dev) (which make heavy use of distinct queries) are about 94% better:

```
$ benchstat old.txt new.txt
name              old time/op  new time/op  delta
_ProfileTypes-10   760ms ± 7%    49ms ± 8%  -93.60%  (p=0.008 n=5+5)
```

This was a pretty extreme benchmark with _lots_ of data. A typical query is now <10ms, and Parca servers only ingesting a single type of profile (eg. only using the [Parca Agent](https://github.com/parca-dev/parca-agent) as it only ingests CPU profiles) even faster.